Clean up handling of locked IOs
Locked IOs can be selected for drag, which displays fail symbols and
prevents any dragging (e.g. no canvas drag).
Locked IOs can be the target of another drag, in which case a fail
symbol is shown and no wire can be drawn.
The fail symbol is now a single element that is repositioned as needed.